Updated Prompts
Still WIP, but idea is in "open saved analysis", options to "Select Folder" with list of folders in /Analysis/ or "Search for Analysis" with type-ins for features of analysis. For "Start new analysis", type-in analysis name, type-ins features, and output list of file(s) that correspond to features, and select to launch plot for it.